data-import/requirements.txt
pandas==2.0.3
influxdb-client==1.36.1
python-dotenv==1.0.0